HypothesisTestsExperimentalRedo

Documentation for HypothesisTestsExperimentalRedo.

HypothesisTestsExperimentalRedo.HypothesisTest

HypothesisTestsExperimentalRedo.NullHypothesis

HypothesisTestsExperimentalRedo.OneSample

HypothesisTestsExperimentalRedo.SummaryStats

HypothesisTestsExperimentalRedo.TTest

HypothesisTestsExperimentalRedo.TwoSample

HypothesisTestsExperimentalRedo.checktest

HypothesisTestsExperimentalRedo.dist_apparent

HypothesisTestsExperimentalRedo.dist_under_H0

HypothesisTestsExperimentalRedo.parameter_of_interest

HypothesisTestsExperimentalRedo.point_est

HypothesisTestsExperimentalRedo.pval

HypothesisTestsExperimentalRedo.reject_H0

HypothesisTestsExperimentalRedo.test_statistic

HypothesisTestsExperimentalRedo.ttest

HypothesisTestsExperimentalRedo.ttest

StatsAPI.confint

StatsAPI.dof

HypothesisTestsExperimentalRedo.HypothesisTest — Type

mutable struct HypothesisTest{T}
    H0::NullHypothesis
    test::T
    data::Vector{SummaryStats}
    level::Real
end

A struct to hold all information required to perform a hypothesis test. The results are computed when needed to ensure that the results never go out of sync with the test input.

HypothesisTestsExperimentalRedo.NullHypothesis — Type

NullHypothesis(POI, operator, value) NullHypothesis(operator, value)

A struct to hold a null hypothesis. POI can be a string or a symbol, and represents the P̲arameter O̲f I̲nterest. It defaults to "POI", but should be set to something to convey information about what the parameter of interest represents, e.g. "μ₀" when the sample mean is being tested.

Fieldnames:

POI: Only important for pretty printing.

operator: Determines if the test is one or two sided, and to which side.

value: The value that is compared with the parameter of interest.

Examples

julia> NullHypothesis(≤, 50)
POI ≤ 50

julia> NullHypothesis(>=, 50)
POI ≥ 50

julia> NullHypothesis("μ₀", >, 50)
μ₀ > 50

julia> NullHypothesis(:μ₀, ==, 50)
μ₀ = 50

HypothesisTestsExperimentalRedo.OneSample — Type

OneSample(paired::Bool)

A test variant that indicates that a single dataset is to be tested.

An instance of OneSample is passed to the relevant test (e.g. TTest). The relevant test is then passed to a HypothesisTest.

The aargument paired determines if the test is interpreted as paired.

HypothesisTestsExperimentalRedo.SummaryStats — Type

SummaryStats(x::Vector{<:Real}) SummaryStats(n, mean, std)

A struct to hold the summary statistics about a sample required to perform statistical hypothesis tests.

Fieldnames:

n: number of samples

mean: the sample mean

std: the sample standard deviation

Examples:

julia> ss = SummaryStats(randn(100))
Summary Statistics for 100 observations: 
------------------------------
Sample mean = -0.01349895434299083
Sample std dev = 1.0247055651499521

julia> ss.n, ss.mean, ss.std
(100, -0.01349895434299083, 1.0247055651499521)

julia> SummaryStats(100, 10, 0.5)
Summary Statistics for 100 observations:
------------------------------
Sample mean = 10
Sample std dev = 0.5

HypothesisTestsExperimentalRedo.TTest — Type

TTest(OneSample(false))
TTest(OneSample(true))
TTest(TwoSample(false))
TTest(TwoSample(true))

A struct used to specify that a T-test should be performed. There are 2 types of T-test: OneSample - for testing a hypothesis regarding the mean of a single dataset TwoSample - for testing a hypothesis regarding the mean difference of two datasets

For a OneSample test, the boolean argument specifies if the test is paired. For a TwoSample test, the boolean argument specifies if equal variance should be assumed.

HypothesisTestsExperimentalRedo.TwoSample — Type

TwoSample(equalvariance::Bool)

A test variant that indicates that the difference between two datasets is to be tested.

An instance of OneSample is passed to the relevant test (e.g. TTest). The relevant test is then passed to a HypothesisTest.

The argument equalvariance determines if equal variance should be assumed.

HypothesisTestsExperimentalRedo.checktest — Method

checktest(ht::HypothesisTest)

A function that takes a HypothesisTest and returns it if all tests are passed. Test are defined for each testtype. This function should be used for the convenience-constructurs, for the public API

HypothesisTestsExperimentalRedo.dist_apparent — Method

dist_apparent(H0::NullHypothesis, testtype::TestType, data::Vector{SummaryStats})
dist_apparent(ht::HypothesisTest)

For any test, compute the distribution indicated by the data samples.

If the test-type is e.g. a t-test, then the apparent distribution will the the scaled and shifted TDist that best corresponds to the data.

HypothesisTestsExperimentalRedo.dist_under_H0 — Method

dist_under_H0(H0::NullHypothesis, testtype::TestType, data::Vector{SummaryStats})
dist_under_H0(ht::HypothesisTest)

For any test, compute the distribution under the null-hypothesis. Used to compared to the point-estimate to get p-values. This package deals only with the actual distributions, and does at no point use normalized ditributions.

HypothesisTestsExperimentalRedo.parameter_of_interest — Method

parameterofinterest(x)

A function to print the parameter of interest in human form for displaying HypothesisTest, e.g. "Mean" for a TTest

HypothesisTestsExperimentalRedo.point_est — Method

    point_est(testtype::TestType, data::Vector{SummaryStats})
    point_est(ht::HypothesisTest)

HypothesisTestsExperimentalRedo.pval — Method

pval(H0::NullHypothesis, dist::Distribution, pointestimate::Real)
pval(ht::HypothesisTest)

HypothesisTestsExperimentalRedo.reject_H0 — Method

reject_H0(ht::HypothesisTest)

Returns trueif the p-value is smaller that 1 minus the confidence level, andfalse` if else.

HypothesisTestsExperimentalRedo.test_statistic — Method

test_statistic(ht::HypothesisTestt

Return the test statistic of a given hypothesis test ht

HypothesisTestsExperimentalRedo.ttest — Function

ttest(obs1::Vector{<:Real}, obs2::Vector{<:Real}, Δμ₀=0; kwargs...)

Test the null-hypothesis that the difference between the observation means (mean(obs1)-mean(obs2)) is equal to Δμ₀, against the alternative hypothesis that it is not.

To perform a one-sided test, set op to one of (>, ≥, ≤ or <).

The keyword argument equalvariance::Bool determines if equal variance between the underlying distributions should be assumed.

HypothesisTestsExperimentalRedo.ttest — Function

ttest(obs::Vector{<:Real}, μ₀=0; kwargs...)
ttest(obs::Vector{<:Real}, op, μ₀=0; kwargs...)

Test the null-hypothesis that the mean of the observations is equal to μ₀, against the alternative hypothesis that it is not.

To perform a one-sided test, set op to one of (>, ≥, ≤ or <).

A paired one-sample t-test is equivalent to a normal one-sample t-test of the elementwise difference between the datasets (dataset1 .- dataset2). To perform such a test, give dataset1 .- dataset2 as the observations, and set the keyword argument paired to true to reflect this in the printing.

StatsAPI.confint — Method

confint(H0::NullHypothesis, dist::Distribution; level::Real)
confint(ht::HypothesisTest)

StatsAPI.dof — Method

dof(ss::SummaryStats) = ss.n-1
dof(data::Vector{SummaryStats}) = sum(dof, data)